requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
SQLAlchemy>=2.0.0
xhtml2pdf>=0.2.11
python-dotenv>=1.0.0
gunicorn>=21.2.0
Flask-Migrate>=4.0.0
Werkzeug>=2.3.0
alembic>=1.12.0
Jinja2>=3.1.0
reportlab>=4.0.0  # Required by xhtml2pdf